+++ inc/simplepdf.class.php
@@ -32,9 +32,7 @@
 
 //use TCPDF;
 
-define('K_PATH_IMAGES', Plugin::getPhpDir('pdf') . '/pics/');
-
-
+define('K_PATH_IMAGES', '');
 class PluginPdfSimplePDF
 {
     // Page orientation
@@ -120,17 +118,21 @@ class PluginPdfSimplePDF
         $this->header = $msg;
         $this->pdf->resetHeaderTemplate();
         $this->pdf->SetTitle($msg);
-        $configurationValues = Config::getConfigurationValues('core', ['version']);
-        $current_version     = $configurationValues['version'];
-        switch ($current_version) {
-            case '0.85.3':
-            case '0.85.4':
-            case '0.85.5':
-                $this->pdf->SetHeaderData('fd_logo.jpg', 15, $msg, '');
-                break;
+        $config = PluginPdfConfig::getInstance();
 
-            default:
-                $this->pdf->SetHeaderData('fd_logo.png', 15, $msg, '');
+        $params = [
+            'entities_id' => Session::getActiveEntity(),
+            'logo'        => '',
+        ];
+        $hook = Plugin::doHookFunction('import_logo', $params);
+        if (
+            !empty($hook['logo_path'])
+            && $config->getField('use_branding_logo')
+        ) {
+            $this->pdf->SetHeaderData($hook['logo_path'], 15, $msg, '');
+        } else {
+            $path = Plugin::getPhpDir('pdf') . '/pics/';
+            $this->pdf->SetHeaderData($path . 'fd_logo.png', 15, $msg, '');
         }
     }